+++ packages/storefront/packages/next/__tests__/apps/pages-router/README.md
@@ -1,0 +1,9 @@
+# Example usage of `@vue-storefront/next` with Pages Router
+
+Please check
+
+- `sdk.config.ts` file for the SDK configuration
+- `hooks/sdk.ts` file for the SDK context creation
+- `pages/_app.tsx` file for the example usage of how to register the SDK context in the root component
+- `pages/ssr.tsx` file for the example usage of how to use the SDK in the Server Side rendered page
+- `pages/csr.tsx` file for the example usage of how to use the SDK in the Client Side rendered page
