+++ packages/storefront/packages/next/__tests__/apps/pages-router/package.json
@@ -1,6 +1,4 @@
 {
-  "name": "pages-router",
-  "version": "0.1.0",
   "private": true,
   "scripts": {
     "dev": "next dev",
