+++ packages/storefront/packages/next/src/sdk/index.tsx
@@ -2,48 +2,59 @@
 /* eslint-disable import/no-extraneous-dependencies */
 /* eslint-disable import/no-unresolved */
 import { buildModule, initSDK } from "@vue-storefront/sdk";
-import { createSdkContext } from "./SdkContext";
+import * as SdkContext from "./SdkContext";
 import { composeMiddlewareUrl } from "./composeMiddlewareUrl";
+import { resolveDynamicContext } from "./resolveDynamicContext";
 import {
   Config,
   CreateSdkOptions,
-  DynamicContext,
+  CreateSdkReturn,
   GetSdkContext,
 } from "./types";
 
 export type { CreateSdkOptions } from "./types";
 
-const BLACKLISTED_HEADERS = ["host"];
-
-function isAppRouterHeaders(
-  headers: ReturnType<GetSdkContext["getRequestHeaders"]>
-): headers is Headers {
-  return headers && "entries" in headers;
-}
-
-function resolveDynamicContext(context: GetSdkContext): DynamicContext {
-  return {
-    ...context,
-    getRequestHeaders() {
-      const headers = context.getRequestHeaders?.() || {};
-      const resolvedHeaders = isAppRouterHeaders(headers)
-        ? Object.fromEntries(headers.entries())
-        : headers;
-      const allowedHeaders = Object.fromEntries(
-        Object.entries(resolvedHeaders).filter(
-          ([key]) => !BLACKLISTED_HEADERS.includes(key)
-        )
-      );
-
-      return allowedHeaders;
-    },
-  };
-}
-
+/**
+ * Creates an SDK for the given configuration definition.
+ * @param options - The options for creating the SDK.
+ * @param configDefinition - The configuration definition for the SDK.
+ * @returns An object containing the `getSdk` function and the `createSdkContext` function.
+ * @example
+ * ```tsx
+ * import {
+ *   contentfulModule,
+ *   ContentfulModuleType,
+ * } from "@vsf-enterprise/contentful-sdk";
+ * import { unifiedModule } from "@vsf-enterprise/unified-sdk";
+ * import { CreateSdkOptions, createSdk } from "@vue-storefront/next";
+ * import type { UnifiedApiExtension } from "../storefront-middleware/middleware.config";
+ *
+ * const options: CreateSdkOptions = {
+ *   middleware: {
+ *     apiUrl: "http://localhost:4000",
+ *   },
+ * };
+ *
+ * export const { getSdk, createSdkContext } = createSdk(
+ *   options,
+ *   ({ buildModule, middlewareUrl, getRequestHeaders }) => ({
+ *     unified: buildModule(unifiedModule<UnifiedApiExtension>, {
+ *       apiUrl: middlewareUrl + "/commerce",
+ *       requestOptions: {
+ *         headers: getRequestHeaders,
+ *       },
+ *     }),
+ *     contentful: buildModule<ContentfulModuleType>(contentfulModule, {
+ *       apiUrl: middlewareUrl + "/cntf",
+ *     }),
+ *   }),
+ * );
+ * ```
+ */
 export function createSdk<TConfig extends Record<string, any>>(
   options: CreateSdkOptions,
-  config: Config<TConfig>
-) {
+  configDefinition: Config<TConfig>
+): CreateSdkReturn<TConfig> {
   function getSdk(dynamicContext: GetSdkContext = {}) {
     const { getRequestHeaders } = resolveDynamicContext(dynamicContext);
     const middlewareUrl = composeMiddlewareUrl({
@@ -51,7 +62,7 @@ export function createSdk<TConfig extends Record<string, any>>(
       headers: getRequestHeaders(),
     });
 
-    const resolvedConfig = config({
+    const resolvedConfig = configDefinition({
       buildModule,
       getRequestHeaders,
       middlewareUrl,
@@ -60,8 +71,12 @@ export function createSdk<TConfig extends Record<string, any>>(
     return initSDK(resolvedConfig);
   }
 
+  function createSdkContext() {
+    return SdkContext.createSdkContext(getSdk());
+  }
+
   return {
     getSdk,
-    createSdkContext: () => createSdkContext(getSdk()),
+    createSdkContext,
   };
 }
